% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedDevices-package.R
\docType{package}
\name{MedDevices-package}
\alias{MedDevices}
\alias{MedDevices-package}
\title{MedDevices: Medical Devices data package}
\description{
MedDevice data package provide an information about medical devices and details of recall. This information provided by U.S.FDA.
}
\author{
\strong{Maintainer}: Sahar Abuzaid \email{saharabuzaid@outlook.com} (\href{https://orcid.org/0000-0003-2635-1696}{ORCID})

}
\keyword{internal}
