% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedDevices.R
\docType{data}
\name{MedDevices}
\alias{MedDevices}
\title{Data of Medical Devices Recalls}
\format{
A tibble with 12423 rows and 14 variables:
\describe{
\item{device_id}{dbl distinction number for each device in the data}
\item{device_name}{chr  name of the device}
\item{device_classification}{chr  denoting the class of the device base on hospital department that the device used in}
\item{manufacturers_name}{chr  name of device's manfacturer}
\item{parent_company}{chr name of manfacturer's parent company}
\item{distributed_to}{chr  denoting the spatial distribution of the device}
\item{implanted}{chr  deonating is the device implantable or not}
\item{risk_class}{chr denoting the class of the device base on risk level}
\item{reason}{chr  denoting the potential danger due to recalled malfunction}
\item{determined_cause}{chr  denoting the cause of the malfunction}
\item{Recall_classification}{chr  denoting the relative degree of risk}
\item{date_initiated_by_firm}{dbl  initiation date of recall by the company}
\item{date_posted}{dbl  date of posting the recall by FDA in database}
\item{date_terminated}{dbl  date of terminating the recall by FDA}
}
}
\source{
\url{https://medicaldevices.icij.org/p/download}
}
\usage{
MedDevices
}
\description{
Medical Devices Recalls provided by The Food and Drug Administration (FDA) in USA for the devices that terminated
}
\keyword{datasets}
